My intuition was correct, phalanx cost elo
With 4000 games of confidence it showed about 10 elo for increasing value of phalanx. The value was only low on the final iteration of tuning so likely just an unlucky coincidence. Still I would at some point like to use SPSA to tune it on self play as a simple for of RL.

Good to see that the -10 elo from the last patch is a product of bad tuning and not the feature simplification.